services/admin/src/requirements.txt
django==5.1.5
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-split-settings==1.3.2
psycopg==3.2.4
psycopg-binary==3.2.4
python-dotenv==1.0.1
uWSGI==2.0.28
